requirements.txt
streamlit==1.31.0
yfinance==0.2.36
plotly==5.18.0
pandas>=2.0.0
requests==2.31.0
python-dotenv==1.0.1
numpy>=1.24.0
pyyaml==6.0.1
alpaca-py==0.21.0
psutil>=5.9.0
